fix: Use `content_type.mime_type` instead of direct header access (#36)
In some environments, retrieving the `content-type` header directly from the
headers hash was returning an empty string. In those environments, the header
was stored as `Content-Type`. By updating the code to use the provided
`content_type` method provided on the response object, we can shield ourselves
from casing issues like this in the future.